Create Patents IP data.

create_patents() creates a data frame with the number of patents filed in a geographic area for a given year, and the average number of backwards citations that a patent application included.

Usage

create_patents(year, geography = "sa2")

Arguments

year

numeric. The year the patents were filed.

geography

string. The geographic region of interest. Defaults to SA2.

Value

A data frame of patents by geography for a given year.

Details

Only patents applied for in Australia are included. This data is derived from the Intellectual Property Government Open Data, and includes:

IPGOD 101 Patents Summary

IPGOD 102 Patents Applicant Information

IPGOD 110 Patents Citation Information

More information about the IPGOD data is available through the IPGOD Data Dictionary.

Examples

create_patents(2016, "SA2")
#> Joining, by = "sa2_name"
#> # A tibble: 1,525 × 4
#>    sa2_name                   year patents backwards_citations
#>    <chr>                     <dbl>   <int>               <dbl>
#>  1 Abbotsford                 2016      68                1396
#>  2 Aberfoyle Park             2016       1                   1
#>  3 Acacia Gardens             2016       1                   1
#>  4 Acton                      2016     189                 998
#>  5 Adelaide                   2016     186                1998
#>  6 Adelaide Hills             2016       4                   5
#>  7 Agnes Water - Miriam Vale  2016       3                   5
#>  8 Ainslie                    2016      91                 700
#>  9 Airlie - Whitsundays       2016       4                   4
#> 10 Airport West               2016       3                   3
#> # … with 1,515 more rows