Industrial Comparative Advantage

ica() creates a data frame of the industrial comparative advantages of Australian regions for 2011 and 2016.

Usage

ica(year, geography = "sa2", adjust = FALSE, ...)

Arguments

year

numeric. The year to calculate the industrial comparative advantage. Must be a Census year (2011, 2016, 2021).

geography

string. The geography level to calculate the industrial comparative advantage. Defaults to SA2.

adjust

logical. TRUE to remove industries classified as "not further defined"

...

Additional options passed to the calculation of the location quotient. See location_quotient for more information.

Value

a data frame of industrial comparative advantages.

Details

The industrial comparative advantage data is derived from employment by industry data from the Australian Bureau of Statistics (ABS) Census based on the place that people work, rather than their place of residence. Industry of employment is measured at the 2006 ANZSIC Subdivision. Employment by industry alone is not an adequate measure of industrial strength, because it is not clear where one region outperforms another, as total employment differs across regions. For example, Preschool and School education is the largest employing industry in more than 25\ in which industries a region employs a higher share than the Australian average.

Census data is randomised in cases where a small number of people work in a specific industry in a region. This is done to avoid the release of confidential data. As such, regions in which fewer than 150 people work, are excluded by default.

See vignette("industrial_growth_opportunities") for more details.

Examples

ica(year = 2016, geography = "sa3")
#> # A tibble: 35,700 × 4
#>    sa3_name      industry                                              ica  year
#>    <chr>         <chr>                                               <dbl> <dbl>
#>  1 Adelaide City Accommodation                                      1.64    2016
#>  2 Adelaide City Accommodation and Food Services, nfd               1.09    2016
#>  3 Adelaide City Administrative Services                            1.20    2016
#>  4 Adelaide City Administrative and Support Services, nfd           3.08    2016
#>  5 Adelaide City Adult, Community and Other Education               0.651   2016
#>  6 Adelaide City Agriculture                                        0.0437  2016
#>  7 Adelaide City Agriculture, Forestry and Fishing Support Services 0.0887  2016
#>  8 Adelaide City Agriculture, Forestry and Fishing, nfd             0.427   2016
#>  9 Adelaide City Air and Space Transport                            0.314   2016
#> 10 Adelaide City Aquaculture                                        0.0696  2016
#> # … with 35,690 more rows
#> # ℹ Use `print(n = ...)` to see more rows

# Remove industries classified as "not further defined"
ica(year = 2016, adjust = TRUE)
#> # A tibble: 195,134 × 4
#>    sa2_name  industry                                             ica  year
#>    <chr>     <chr>                                              <dbl> <dbl>
#>  1 Braidwood Agriculture                                        10.1   2016
#>  2 Braidwood Aquaculture                                         0     2016
#>  3 Braidwood Forestry and Logging                                8.43  2016
#>  4 Braidwood Fishing, Hunting and Trapping                       0     2016
#>  5 Braidwood Agriculture, Forestry and Fishing Support Services  9.84  2016
#>  6 Braidwood Coal Mining                                         0     2016
#>  7 Braidwood Oil and Gas Extraction                              0     2016
#>  8 Braidwood Metal Ore Mining                                    0     2016
#>  9 Braidwood Non-Metallic Mineral Mining and Quarrying           5.20  2016
#> 10 Braidwood Exploration and Other Mining Support Services       0     2016
#> # … with 195,124 more rows
#> # ℹ Use `print(n = ...)` to see more rows

# Increase the minimum regional employment
ica(year = 2016, min_value = 1000)
#> # A tibble: 238,245 × 4
#>    sa2_name  industry                                             ica  year
#>    <chr>     <chr>                                              <dbl> <dbl>
#>  1 Braidwood Agriculture, Forestry and Fishing, nfd             25.6   2016
#>  2 Braidwood Agriculture                                         9.89  2016
#>  3 Braidwood Aquaculture                                         0     2016
#>  4 Braidwood Forestry and Logging                                8.53  2016
#>  5 Braidwood Fishing, Hunting and Trapping                       0     2016
#>  6 Braidwood Agriculture, Forestry and Fishing Support Services  9.97  2016
#>  7 Braidwood Mining, nfd                                         0     2016
#>  8 Braidwood Coal Mining                                         0     2016
#>  9 Braidwood Oil and Gas Extraction                              0     2016
#> 10 Braidwood Metal Ore Mining                                    0     2016
#> # … with 238,235 more rows
#> # ℹ Use `print(n = ...)` to see more rows

# Remove industries which do not meet a minimum employment threshold
ica(year = 2016, total_var = "industry_employment", min_value = 2000)
#> # A tibble: 224,631 × 4
#>    sa2_name  industry                                             ica  year
#>    <chr>     <chr>                                              <dbl> <dbl>
#>  1 Braidwood Agriculture, Forestry and Fishing, nfd             25.6   2016
#>  2 Braidwood Agriculture                                         9.88  2016
#>  3 Braidwood Aquaculture                                         0     2016
#>  4 Braidwood Forestry and Logging                                8.52  2016
#>  5 Braidwood Fishing, Hunting and Trapping                       0     2016
#>  6 Braidwood Agriculture, Forestry and Fishing Support Services  9.96  2016
#>  7 Braidwood Mining, nfd                                         0     2016
#>  8 Braidwood Coal Mining                                         0     2016
#>  9 Braidwood Oil and Gas Extraction                              0     2016
#> 10 Braidwood Metal Ore Mining                                    0     2016
#> # … with 224,621 more rows
#> # ℹ Use `print(n = ...)` to see more rows