Industrial Growth Opportunities

igo() creates a data frame of the industrial growth opportunities across Statistical Area Level 2 regions for 2011 and 2016.

Usage

igo(
  year,
  region = NULL,
  product = NULL,
  .export_value_limit = 0,
  .cog_limit = 0,
  .rca_limit = 1,
  .ica_limit = 1,
  ...
)

Arguments

year

numeric. Which year to calculate the industrial growth opportunities.

region

character. Show only industrial growth opportunities for a given region. The default (NULL) will return industrial growth opportunities for all regions.

product

character. Show only regions with an opportunity in a given product. The default (NULL) will return industrial growth opportunities for all products

.export_value_limit

(optional) numeric. The minimum state export value for a product required to be considered a growth opportunity for a region. The default (0) excludes products which were not exported.

.cog_limit

(optional) numeric. The minimum benefit that a product will generate for a region. The default (0) excludes all products with negative benefits.

.rca_limit

(optional) numeric. The maximum value for the level of revealed comparative advantage in the production of an opportunity in a state. The default (1) excludes all products which are already produced with comparative advantage.

.ica_limit

(optional) numeric. The minimum value for the level of industrial comparative advantage. The default (1) excludes all products which are made in industries in which the region does not have a higher proportion of employment than Australia.

...

(optional). Additional arguments passed to ica().

Value

a data frame of industrial growth opportunities.

Details

The industrial growth opportunities are derived from the industrial comparative advantage data (see ica()), and a State based model of economic complexity, developed by the Australian Industrial Transformation Institute at Flinders University, in collaboration with the Government of South Australia.

Industrial growth opportunities capture, at the product level, what industrial development would be:

most beneficial for a region, and

most suitable for a region based on its industrial strengths.

Examples

igo(2016)
#> # A tibble: 317,400 × 17
#>    locat…¹ hs_pr…² expor…³  year diver…⁴ ubiqu…⁵   mcp   eci   pci density   coi
#>    <chr>   <chr>     <dbl> <dbl>   <dbl>   <dbl> <dbl> <dbl> <dbl>   <dbl> <dbl>
#>  1 New So… 0105     717553  2016     178      31     0 0.333  1.06   0.171 0.731
#>  2 New So… 0105     717553  2016     178      31     0 0.333  1.06   0.171 0.731
#>  3 New So… 0105     717553  2016     178      31     0 0.333  1.06   0.171 0.731
#>  4 New So… 0105     717553  2016     178      31     0 0.333  1.06   0.171 0.731
#>  5 New So… 0105     717553  2016     178      31     0 0.333  1.06   0.171 0.731
#>  6 New So… 0105     717553  2016     178      31     0 0.333  1.06   0.171 0.731
#>  7 New So… 0105     717553  2016     178      31     0 0.333  1.06   0.171 0.731
#>  8 New So… 0105     717553  2016     178      31     0 0.333  1.06   0.171 0.731
#>  9 New So… 0105     717553  2016     178      31     0 0.333  1.06   0.171 0.731
#> 10 New So… 0105     717553  2016     178      31     0 0.333  1.06   0.171 0.731
#> # … with 317,390 more rows, 6 more variables: cog <dbl>, rca <dbl>,
#> #   anzsic_subdivision <chr>, sa2_name <chr>, ica <dbl>,
#> #   hs_product_name_short_en <chr>, and abbreviated variable names
#> #   ¹​location_code, ²​hs_product_code, ³​export_value, ⁴​diversity, ⁵​ubiquity

# Industrial growth opportunities for a specific region
igo(2016, region = "Adelaide")
#> # A tibble: 3 × 17
#>   location_code hs_pr…¹ expor…²  year diver…³ ubiqu…⁴   mcp    eci   pci density
#>   <chr>         <chr>     <dbl> <dbl>   <dbl>   <dbl> <dbl>  <dbl> <dbl>   <dbl>
#> 1 South Austra… 4901    2310897  2016     102      37     0 -0.109  1.03  0.105 
#> 2 South Austra… 4902       1794  2016     102      30     0 -0.109  2.19  0.0950
#> 3 South Austra… 4903      12424  2016     102      14     0 -0.109  1.43  0.0646
#> # … with 7 more variables: coi <dbl>, cog <dbl>, rca <dbl>,
#> #   anzsic_subdivision <chr>, sa2_name <chr>, ica <dbl>,
#> #   hs_product_name_short_en <chr>, and abbreviated variable names
#> #   ¹​hs_product_code, ²​export_value, ³​diversity, ⁴​ubiquity

# Specifying a product may help to identify a region best suited to develop a new industry
igo(2016, product = "Artificial graphite")
#> # A tibble: 745 × 17
#>    locat…¹ hs_pr…² expor…³  year diver…⁴ ubiqu…⁵   mcp   eci   pci density   coi
#>    <chr>   <chr>     <dbl> <dbl>   <dbl>   <dbl> <dbl> <dbl> <dbl>   <dbl> <dbl>
#>  1 New So… 3801      25946  2016     178      15     0 0.333  2.12   0.124 0.731
#>  2 New So… 3801      25946  2016     178      15     0 0.333  2.12   0.124 0.731
#>  3 New So… 3801      25946  2016     178      15     0 0.333  2.12   0.124 0.731
#>  4 New So… 3801      25946  2016     178      15     0 0.333  2.12   0.124 0.731
#>  5 New So… 3801      25946  2016     178      15     0 0.333  2.12   0.124 0.731
#>  6 New So… 3801      25946  2016     178      15     0 0.333  2.12   0.124 0.731
#>  7 New So… 3801      25946  2016     178      15     0 0.333  2.12   0.124 0.731
#>  8 New So… 3801      25946  2016     178      15     0 0.333  2.12   0.124 0.731
#>  9 New So… 3801      25946  2016     178      15     0 0.333  2.12   0.124 0.731
#> 10 New So… 3801      25946  2016     178      15     0 0.333  2.12   0.124 0.731
#> # … with 735 more rows, 6 more variables: cog <dbl>, rca <dbl>,
#> #   anzsic_subdivision <chr>, sa2_name <chr>, ica <dbl>,
#> #   hs_product_name_short_en <chr>, and abbreviated variable names
#> #   ¹​location_code, ²​hs_product_code, ³​export_value, ⁴​diversity, ⁵​ubiquity

# Override defaults.
# Increasing the export value limit (from 0) focuses the opportunities
# on those with a stronger existing industrial base.

# Increasing the COG limit (from 0) focuses the opportunities
# on those which would bring a higher benefit to the region.

# Decreasing the RCA limit (from 1) focuses the opportunities
# on those which are less developed in the State.

# Increasing the ICA limit (from 1) focuses the opportunities
# on industries in which a region has a significant comparative advantage.

igo(2016, .export_value_limit = 1000, .cog_limit = 0.5, .rca_limit = 0.5, .ica_limit = 0.5)
#> # A tibble: 317,708 × 17
#>    locat…¹ hs_pr…² expor…³  year diver…⁴ ubiqu…⁵   mcp   eci   pci density   coi
#>    <chr>   <chr>     <dbl> <dbl>   <dbl>   <dbl> <dbl> <dbl> <dbl>   <dbl> <dbl>
#>  1 New So… 0105     717553  2016     178      31     0 0.333  1.06   0.171 0.731
#>  2 New So… 0105     717553  2016     178      31     0 0.333  1.06   0.171 0.731
#>  3 New So… 0105     717553  2016     178      31     0 0.333  1.06   0.171 0.731
#>  4 New So… 0105     717553  2016     178      31     0 0.333  1.06   0.171 0.731
#>  5 New So… 0105     717553  2016     178      31     0 0.333  1.06   0.171 0.731
#>  6 New So… 0105     717553  2016     178      31     0 0.333  1.06   0.171 0.731
#>  7 New So… 0105     717553  2016     178      31     0 0.333  1.06   0.171 0.731
#>  8 New So… 0105     717553  2016     178      31     0 0.333  1.06   0.171 0.731
#>  9 New So… 0105     717553  2016     178      31     0 0.333  1.06   0.171 0.731
#> 10 New So… 0105     717553  2016     178      31     0 0.333  1.06   0.171 0.731
#> # … with 317,698 more rows, 6 more variables: cog <dbl>, rca <dbl>,
#> #   anzsic_subdivision <chr>, sa2_name <chr>, ica <dbl>,
#> #   hs_product_name_short_en <chr>, and abbreviated variable names
#> #   ¹​location_code, ²​hs_product_code, ³​export_value, ⁴​diversity, ⁵​ubiquity