Innovation metrics of SA2 regions for 2011 and 2016.

A dataset containing a measure of innovation, inputs to innovation, and other innovation metrics for SA2 regions.

Usage

regional_innovation

Format

A data frame with 4325 rows and 20 variables:

sa2_name

Name of the SA2 region. For 2011, uses the 2011 Australian Statistical Geography Standard. For 2016, uses the 2016 Australian Statistical Geography Standard

sa2_code

8-digit code assigned to the sa2 name

year

patents

Number of patents per 1000 workers. Patent data is derived from the 2019 Intellectual Property Government Open Data (IPGOD) tables 101 and 102. All patent applications from a valid SA2 are included.

backwards_citations

Number of backward citations. Patent data is derived from the 2019 IPGOD tables 110.

designs

Number of designs per 1000 workers. Designs data is derived from the 2019 IPGOD tables 301 and 302. Only designs whose application originated from a valid SA2 in Australia are included.

trademarks

Number of trademarks per 1000 workers. Trademarks data is derived from the 2019 IPGOD tables 201 and 201. Only trademarks whose application originated from a valid SA2 in Australia are included.

plants

Number of plant breeder rights applications per 1000 workers. Plant Breeder Rights data is derived from the 2019 IPGOD tables 401 and 402. Only plant breeder rights lodged from a valid SA2 in Australia are included.

kibs

Proportion of workers in knowledge intensive business services. See ?kibs_industry_class.

skill

Average skill level of region occupations. Occupation skill levels are derived from the Australian Skills Commission Core Competencies of occupations as defined by the 2006 ANZSCO at the unit level. Each occupation is scored across 10 competencies, receiving a score between 1 and 10 where 10 is the most skilled. The skill level of a region is the weighted average of an occupations skill level, and the share of employment in that occupation in an SA2. Skill ranges from 1 to 10 where 10 is the most skilled.

owner_managers

Proportion of workers who are owner managers of incorporated and unincorporated enterprises. Data is derived from the ABS Census of Population and Housing SIEMP for 2016 and EMTP for 2011.

unis

Number of universities located within SA2 boundaries. University point data is sourced from the AURIN Portal.

tafes

Number of TAFEs located within SA2 boundaries. TAFE point data is sourced from the AURIN Portal.

qualification

Average education level of those employed in the region. Data is sourced from the ABS Census of Population and Housing QALLP for 2011 and 2016. A region is scored based on the proportion of workers within each education qualification level, where Certificate level qualifications are worth 1 and Postgraduate Degree level qualifications are worth 4. The Graduate Diploma and Graduate Certificate level is combined with the Bachelor Degree level. As such, each region receives a score between 1 and 4, where a higher score indicates a higher average level of education.

stem

Proportion of workers with a qualification in a stem field. STEM qualifications refer to any non-school qualifications in Natural and Physical Sciences, Information Technology, Engineering and Related Technologies, and Agriculture, Environmental and Related Studies

employment

Total employment, measured by the ABS 2011 and 2016 Census, Counting Place of Work.

infrastructure

Number of universities and tafes

human_knowledge

Factor derived from kibs, skill, and qualification

patent_output

Factor derived from patents and backwards citations

innovation

Second-order factor derived from human_knowledge and patent_output